@@ -70,7 +70,7 @@ const routes: Routes = [
 @NgModule({
   imports: [
     RouterModule.forRoot(routes, {
-      initialNavigation: 'enabled'
+      initialNavigation: 'enabledBlocking'
     })
   ],
   exports: [RouterModule]
